+++ packages/plugin-video/web/src/tiptap/component.tsx
@@ -2,14 +2,18 @@ import React from 'react';
 import { Component as VideoComponent } from '../video-component';
 import { VIDEO_TYPE } from '../types';
 import { PluginProps } from 'wix-rich-content-editor-common';
+import { convertBlockDataToRicos } from 'ricos-content/libs/migrateSchema';
 
-export const Video: React.FC<PluginProps> = ({ context, componentData }) => {
+export const Video: React.FC<PluginProps> = ({ context, componentData, updateAttributes }) => {
   const { theme, t, config = {} } = context;
   const settings = config[VIDEO_TYPE] || {};
   const blockProps = {
     setFocusToBlock: () => null,
   };
   const setComponentUrl = () => null;
+  const store = {
+    update: (type, data) => updateAttributes(convertBlockDataToRicos(VIDEO_TYPE, data)),
+  };
 
   return (
     <VideoComponent
@@ -20,6 +24,7 @@ export const Video: React.FC<PluginProps> = ({ context, componentData }) => {
       blockProps={blockProps}
       setComponentUrl={setComponentUrl}
       onClick={() => {}}
+      store={store}
     />
   );
 };